Created CustomDate class
